add processor and middleend

+++ headers/frontend_h/lexer.h
@@ -69,4 +69,4 @@ int symbol_table_add(const char * name);
 int symbol_table_resize(SymbolTable *sb);
 int symbol_table_get_or_add(const char * name);
 void symbol_table_destroy(SymbolTable * sb);
-const char * get_id_name(int id_index);
+const char * get_id_name(size_t id_index);
